Androidのサウンドシステムもvgs2sound.cに統合
